chore: rename and align agent name for coagents-wait-user-input
