+++ examples/coagents-wait-user-input/README.md
@@ -64,4 +64,4 @@ Make sure to create teh `.env` mentioned above first!
 A few things to try if you are running into trouble:
 
 1. Make sure there is no other local application server running on the 8000 port.
-2. Under `/agent/my_agent/demo.py`, change `0.0.0.0` to `127.0.0.1` or to `localhost`
+2. Under `/agent/weather_agent/demo.py`, change `0.0.0.0` to `127.0.0.1` or to `localhost`
