+++ examples/coagents-wait-user-input/agent/weather_agent/agent.py
